Adds org.simple.json project source files into project
The JSONDoclet requires all classes to be present at the output directory "bin", however
as this project is a plugin and the json project is a normal Java project. It is tedious
and error-prone to be automatically configure this to work everytime. So the decision was
to include the Json project source files into this project.

Anybody who uses this project must refer to the Json project licencing and copyright
information available at their website http://code.google.com/p/json-simple/